Minor Updates in CallLog.

+++ Pegasus.API/Models/Data/CallLog.cs
@@ -4,8 +4,8 @@ namespace Pegasus.API.Models.Data;
 
 public class CallLog
 {
-    public int Id { get; set; }
-    public int? ContactId { get; set; }
+    public long Id { get; set; }
+    public long? ContactId { get; set; }
     public CallType Type { get; set; }
     public DateTime CallDate { get; set; }
     public int Duration { get; set; }
